+++ scripts/scripts.js
@@ -13,6 +13,8 @@ import {
   loadCSS,
 } from './lib-franklin.js';
 
+import integrateMartech from './third-party.js';
+
 const LCP_BLOCKS = ['hero']; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'project-1'; // add your RUM generation information here
 
@@ -113,6 +115,14 @@ export function addFavIcon(href) {
   }
 }
 
+function initPartytown() {
+  window.partytown = {
+    lib: '/scripts/partytown/',
+    forward: ['dataLayer.push'],
+  };
+  import('./partytown/partytown.js');
+}
+
 /**
  * loads everything that doesn't need to be delayed.
  */
@@ -132,6 +142,9 @@ async function loadLazy(doc) {
   sampleRUM('lazy');
   sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
   sampleRUM.observe(main.querySelectorAll('picture > img'));
+
+  integrateMartech();
+  initPartytown();
 }
 
 /**
