+++ scripts/third-party.js
@@ -1,0 +1,16 @@
+//  Load Google Tag Manager
+const fireGTM = (w, d, s, l, i) => {
+  w[l] = w[l] || [];
+  w[l].push({ 'gtm.start': new Date().getTime(), event: 'gtm.js' });
+  const f = d.getElementsByTagName(s)[0];
+  const j = d.createElement(s);
+  j.type = 'text/partytown';
+  const dl = l !== 'dataLayer' ? `&l=${l}` : '';
+  j.async = true;
+  j.src = `https://www.googletagmanager.com/gtm.js?id=${i}${dl}`;
+  f.parentNode.insertBefore(j, f);
+};
+
+export default function integrateMartech() {
+  fireGTM(window, document, 'script', 'dataLayer', 'GTM-5V9FN8V');
+}
